Move all classes back into single module to make it easier
