+++ examples/composer.html
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html ng-app="composer">
     <head>
-        <title>full example</title>
+        <title>composer example</title>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="ol.css" type="text/css">
