+++ examples/viewer.html
@@ -1,15 +1,13 @@
 <!DOCTYPE html>
 <html ng-app="viewer">
     <head>
-        <title>full example</title>
+        <title>viewer example</title>
         <meta charset="UTF-8">
         <meta name="viewport" content="width=device-width, initial-scale=1.0">
         <link rel="stylesheet" href="ol.css" type="text/css">
         <link rel="stylesheet" href="../bower_components/bootstrap/dist/css/bootstrap.min.css" type="text/css">
         <link rel="stylesheet" href="../node_modules/vis/dist/vis.min.css" type="text/css"/>
         <link rel="stylesheet" href="../bower_components/nouislider/distribute/jquery.nouislider.min.css" type="text/css">
-        <link rel="stylesheet" href="../dist/story-tools-edit.css" type="text/css">
-        <link rel="stylesheet" href="../bower_components/angular-bootstrap-colorpicker/css/colorpicker.css" type="text/css">
 
         <script src="../bower_components/angular/angular.min.js" ></script>
         <script src="../bower_components/jquery/dist/jquery.min.js"></script>
@@ -18,11 +16,8 @@
         <script src="../bower_components/angular-bootstrap-colorpicker/js/bootstrap-colorpicker-module.js"></script>
         <script src="ol.js" ></script>
         <script src="../dist/story-tools-core-tpls.js"></script>
-<!--        <script src="../dist/story-tools-edit-tpls.js"></script>-->
         <script src="../dist/story-tools-core.js"></script>
-<!--        <script src="../dist/story-tools-edit.js"></script>-->
         <script src="../dist/story-tools-core-ng.js"></script>
-<!--        <script src="../dist/story-tools-edit-ng.js"></script>-->
         <script src="../dist/ows.js"></script>
         <script src="../dist/story-tools-mapstory.js"></script>
         <script  src="common-viewer.js"></script>
